requirements.txt
pythainlp==4.0.2
datasets==2.18.0
accelerate==0.27.2
faiss-cpu
sentence-transformers==2.5.1
python-crfsuite==0.9.10
numpy==1.26.2
pandas==2.1.3
torch==2.1.2
transformers==4.35.2
gensim==4.3.2
safetensors==0.4.2
scikit-learn==1.2.2
openpyxl==3.1.2
gradio==4.21.0
evaluate==0.4.1
rouge-score==0.1.2
bert-score==0.3.13
optuna==3.5.0
matplotlib==3.8.3